feat: Initialize SQLAlchemy ORM and refactor database models
- Update requirements.txt to include SQLAlchemy for ORM integration
- Create src/database.py to manage SQLAlchemy engine, sessions, and SQLite database path (panda_ledger.db)
- Define core SQLAlchemy ORM models in src/models.py (Account, Category, Transaction) with enums and relationships
- Extract payroll logic (PayrollCalculator, TaxResult, PaycheckResult) from models.py into a new src/payroll.py file
- Update src/main.py imports to use the new payroll.py and database.py modules
- Modify main.py to initialize both the new SQLAlchemy tables and legacy SQLite tables to ensure UI continuity
- Centralize database path logic in database.py to securely utilize the OS-specific AppData directory
